@@ -8,8 +8,9 @@ authors = [
     {name = "Ashley Whetter", email = "ashley@awhetter.co.uk"},
 ]
 readme = "README.rst"
+license = {file = "LICENSE.rst"}
 classifiers = [
-    "Development Status :: 4 - Beta",
+    "Development Status :: 5 - Production/Stable",
     "Environment :: Plugins",
     "Framework :: Sphinx :: Extension",
     "Intended Audience :: Developers",
@@ -17,31 +18,35 @@ classifiers = [
     "Natural Language :: English",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Documentation :: Sphinx",
 ]
-requires-python = ">=3.8"
-dynamic = ["version", "description"]
+requires-python = ">=3.9"
 dependencies = [
     "graphqldomain>=0.2.0",
     "myst-parser>=0.16.0",
     "sphinx>=4.0",
     "typing-extensions>=4.0",
 ]
+dynamic = ["description", "version"]
 
 [project.urls]
 Source = "https://github.com/AWhetter/autogqlschema"
 Documentation = "https://autogqlschema.readthedocs.org"
 
-[tool.pylint]
-disable = "R,unused-argument"
+[tool.flit.sdist]
+include = ["tests/"]
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
 
 [tool.towncrier]
 directory = "doc/changes"
 filename = "CHANGELOG.rst"
 package = "autogqlschema"
-title_format = "autogqlschema v{version} ({project_date})"
+title_format = "v{version} ({project_date})"
+underlines = ["-", "^", "\""]
