@@ -6,45 +6,44 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'autogqlschema'
-copyright = '2023, Ashley Whetter'
-author = 'Ashley Whetter'
-release = '0.1.0'
+project = "autogqlschema"
+copyright = "2023, Ashley Whetter"
+author = "Ashley Whetter"
+release = "0.1.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
-    'autogqlschema',
-    'sphinx.ext.intersphinx',
+    "autogqlschema",
+    "sphinx.ext.intersphinx",
 ]
 
-templates_path = ['_templates']
+templates_path = ["_templates"]
 exclude_patterns = []
-
 
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'furo'
-html_static_path = ['_static']
+html_theme = "furo"
+html_static_path = ["_static"]
 
 
 # -- Options for intersphinx -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html#configuration
 
 intersphinx_mapping = {
-    'graphqldomain': ('https://graphqldomain.readthedocs.io/en/latest/', None),
-    'sphinx': ('https://www.sphinx-doc.org/en/master', None),
-    'myst': ('https://myst-parser.readthedocs.io/en/latest/', None),
+    "graphqldomain": ("https://graphqldomain.readthedocs.io/en/latest/", None),
+    "sphinx": ("https://www.sphinx-doc.org/en/master", None),
+    "myst": ("https://myst-parser.readthedocs.io/en/latest/", None),
 }
 
 
 def setup(app):
     app.add_object_type(
-        'confval',
-        'confval',
-        objname='configuration value',
-        indextemplate='pair: %s; configuration value',
+        "confval",
+        "confval",
+        objname="configuration value",
+        indextemplate="pair: %s; configuration value",
     )
